// src/core/domain/enums/order-status.enum.ts
export enum OrderStatus {
  CREATED = "CREATED",
  PENDING_APPROVAL = "PENDING_APPROVAL",
  PENDING_CONFIRMATION = "PENDING_CONFIRMATION",
  CONFIRMED = "CONFIRMED",
  REJECTED = "REJECTED",
  IN_ADJUSTMENT = "IN_ADJUSTMENT",
  IN_PREPARATION = "IN_PREPARATION",
  IN_TRANSIT = "IN_TRANSIT",
  DELIVERED = "DELIVERED",
  CANCELLED = "CANCELLED",
}